add new

+++ src/main/java/com/codegym/repository/IGeneralRepository.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface IGeneralRepository<E> {
     List<E> findAll();
     void add (E e);
-    E findById(int id);
-    void remove(int id);
+    E findById(long id);
+    void remove(Long id);
     void editUsers(int id,E e);
 
 }
