+++ src/main/java/com/codegym/service/IGeneralService.java
@@ -5,8 +5,8 @@ import java.util.List;
 public interface IGeneralService<E> {
     List<E> findAll();
     void add(E e);
-    E findById(int id);
-    void remove(int id);
+    E findById(long id);
+    void remove(Long id);
     void editUsers(int id,E e);
 }
