+++ src/main/java/com/codegym/service/impl/UserServiceImpl.java
@@ -1,9 +1,7 @@
 package com.codegym.service.impl;
 
 import com.codegym.model.Users;
-import com.codegym.model.UsersForm;
 import com.codegym.repository.IUserRepository;
-import com.codegym.repository.impl.UserRepositoryImpl;
 import com.codegym.service.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -25,13 +23,13 @@ public class UserServiceImpl implements IUserService {
     }
 
     @Override
-    public Users findById(int id) {
+    public Users findById(long id) {
         Users users = userRepository.findById(id);
         return users;
     }
 
     @Override
-    public void remove(int id) {
+    public void remove(Long id) {
         userRepository.remove(id);
 
     }
